Fixed makefile for local compilation. also now with some basic implementation of kernels.
git-svn-id: svn+ssh://njoubert.com/var/svn@1242 997b1fe7-5dcb-cc6a-e049-b7da040a1fbe

+++ CS149_PA4/ImageCleaner.cu
@@ -135,7 +135,7 @@ __host__ float filterImage(float *real_image, float *imag_image, int size_x, int
 
 
   dim3 dimBlock(BLOCK_SIZE);
-  dim3 dimGrid(ceil((float)SIZEX/BLOCK_SIZE),SIZEY);
+  dim3 dimGrid((int)ceil((float)SIZEX/BLOCK_SIZE),SIZEY);
   printf("  Launching kernel with %d threads per block arranged in a grid of %dx%d.\n", dimBlock.x, dimGrid.x, dimGrid.y);
   gpu_fftx<<<dimGrid, dimBlock>>>(device_real,device_imag,size_x,size_y);
